Revert "Merge main branch - pull request #784 from arawa/fix/concate-for-strings"
This reverts commit 750a2933e2abf192c4ea997bd7c81b1b6d079a9c.

+++ src/SpaceDetails.vue
@@ -1,23 +1,23 @@
 <!--
-	@copyright Copyright (c) 2017 Arawa
+  @copyright Copyright (c) 2017 Arawa
 
-	@author 2021 Baptiste Fotia <baptiste.fotia@arawa.fr>
-	@author 2021 Cyrille Bollu <cyrille@bollu.be>
+  @author 2021 Baptiste Fotia <baptiste.fotia@arawa.fr>
+  @author 2021 Cyrille Bollu <cyrille@bollu.be>
 
-	@license GNU AGPL version 3 or any later version
+  @license GNU AGPL version 3 or any later version
 
-	This program is free software: you can redistribute it and/or modify
-	it under the terms of the GNU Affero General Public License as
-	published by the Free Software Foundation, either version 3 of the
-	License, or (at your option) any later version.
+  This program is free software: you can redistribute it and/or modify
+  it under the terms of the GNU Affero General Public License as
+  published by the Free Software Foundation, either version 3 of the
+  License, or (at your option) any later version.
 
-	This program is distributed in the hope that it will be useful,
-	but WITHOUT ANY WARRANTY; without even the implied warranty of
-	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-	GNU Affero General Public License for more details.
+  This program is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  GNU Affero General Public License for more details.
 
-	You should have received a copy of the GNU Affero General Public License
-	along with this program.  If not, see <http://www.gnu.org/licenses/>.
+  You should have received a copy of the GNU Affero General Public License
+  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 -->
 
 <template>
@@ -75,7 +75,7 @@
 					</NcActionInput>
 					<NcActionButton icon="icon-delete"
 						:close-after-click="true"
-						@click="toggleShowDelWorkspaceModal">
+						@click="deleteSpace">
 						{{ t('workspace', 'Delete space') }}
 					</NcActionButton>
 				</NcActions>
@@ -85,12 +85,6 @@
 		<NcModal v-if="showSelectUsersModal"
 			@close="toggleShowSelectUsersModal">
 			<SelectUsers :space-name="$route.params.space" @close="toggleShowSelectUsersModal" />
-		</NcModal>
-		<NcModal v-if="showDelWorkspaceModal"
-      style="min-heigth: 8rem;"
-      size="small"
-		  @close="toggleShowDelWorkspaceModal">
-		  <RemoveSpace :space-name="$route.params.space" @handle-cancel="toggleShowDelWorkspaceModal" @handle-delete="deleteSpace" />
 		</NcModal>
 	</div>
 </template>
@@ -105,9 +99,8 @@ import NcColorPicker from '@nextcloud/vue/dist/Components/NcColorPicker.js'
 import NcMultiselect from '@nextcloud/vue/dist/Components/NcMultiselect.js'
 import NcModal from '@nextcloud/vue/dist/Components/NcModal.js'
 import SelectUsers from './SelectUsers.vue'
-import RemoveSpace from './RemoveSpace.vue'
 import UserTable from './UserTable.vue'
-import { destroy, rename, checkGroupfolderNameExist } from './services/groupfoldersService.js'
+import { destroy, rename } from './services/groupfoldersService.js'
 
 export default {
 	name: 'SpaceDetails',
@@ -119,7 +112,6 @@ export default {
 		NcModal,
 		NcMultiselect,
 		SelectUsers,
-		RemoveSpace,
 		UserTable,
 	},
 	data() {
@@ -127,7 +119,6 @@ export default {
 			createGroup: false, // true to display 'Create Group' ActionInput
 			renameSpace: false, // true to display 'Rename space' ActionInput
 			showSelectUsersModal: false, // true to display user selection Modal windows
-			showDelWorkspaceModal: false,
 			isESR: false,
 		}
 	},
@@ -147,17 +138,22 @@ export default {
 		// Deletes a space
 		deleteSpace() {
 			const space = this.$route.params.space
-			destroy(this.$store.state.spaces[space])
-				.then(resp => {
-					if (resp.http.statuscode === 200) {
-						this.$store.dispatch('removeSpace', {
-							space: this.$store.state.spaces[space],
-						})
-						this.$router.push({
-							path: '/',
-						})
-					}
-				})
+
+			const isDeleted = window.confirm(t('workspace', 'Are you sure you want to delete the {space} space ?', { space }))
+
+			if (isDeleted) {
+				destroy(this.$store.state.spaces[space])
+					.then(resp => {
+						if (resp.http.statuscode === 200) {
+							this.$store.dispatch('removeSpace', {
+								space: this.$store.state.spaces[space],
+							})
+							this.$router.push({
+								path: '/',
+							})
+						}
+					})
+			}
 		},
 		onNewGroup(e) {
 			// Hides ActionInput
@@ -170,9 +166,9 @@ export default {
 			}
 
 			// Creates group
-			this.$store.dispatch('createGroup', { name: this.$route.params.space, gid, vueInstance: this })
+			this.$store.dispatch('createGroup', { name: this.$route.params.space, gid })
 		},
-		async onSpaceRename(e) {
+		onSpaceRename(e) {
 			// Hides ActionInput
 			this.toggleRenameSpace()
 
@@ -187,45 +183,47 @@ export default {
 					duration: 6000,
 				})
 			}
-
-			const newSpaceName = e.target[0].value
-
-			await checkGroupfolderNameExist(newSpaceName, this)
-
 			// TODO: Change : the key from $root.spaces, groupnames, change the route into new spacename because
 			// the path is `https://instance-nc/apps/workspace/workspace/Aang`
 			const oldSpaceName = this.$route.params.space
-			let responseRename = await rename(this.$store.state.spaces[oldSpaceName], newSpaceName, this)
-			responseRename = responseRename.data
-
-			if (responseRename.statuscode === 204) {
-				const space = { ...this.$store.state.spaces[oldSpaceName] }
-				space.name = responseRename.space
-				space.groups = responseRename.groups
-				this.$store.dispatch('updateSpace', {
-					space,
+			rename(this.$store.state.spaces[oldSpaceName], e.target[0].value)
+				.then(resp => {
+					const data = resp.data
+					if (data.statuscode === 409) {
+						this.$notify({
+							title: t('workspace', 'Error to rename space'),
+							text: t('workspace', data.message),
+							type: 'error',
+							duration: 6000,
+						})
+					}
+					if (data.statuscode === 204) {
+						const space = { ...this.$store.state.spaces[oldSpaceName] }
+						space.name = data.space
+						space.groups = data.groups
+						this.$store.dispatch('updateSpace', {
+							space,
+						})
+						this.$store.dispatch('removeSpace', {
+							space: this.$store.state.spaces[oldSpaceName],
+						})
+						this.$router.push({
+							path: `/workspace/${space.name}`,
+						})
+					}
+					if (data.statuscode === 401) {
+						// TODO: May be to print an error message temporary
+						console.error(data.message)
+					}
+					if (data.statuscode === 400) {
+						this.$notify({
+							title: t('workspace', 'Error to rename space'),
+							text: t('workspace', 'Your Workspace name must not contain the following characters: [ ~ < > { } | ; . : , ! ? \' @ # $ + ( ) % \\\\ ^ = / & * ]'),
+							type: 'error',
+							duration: 6000,
+						})
+					}
 				})
-				this.$store.dispatch('removeSpace', {
-					space: this.$store.state.spaces[oldSpaceName],
-				})
-				this.$router.push({
-					path: `/workspace/${space.name}`,
-				})
-			}
-
-			if (responseRename.statuscode === 401) {
-				// TODO: May be to print an error message temporary
-				console.error(responseRename.message)
-			}
-
-			if (responseRename.statuscode === 400) {
-				this.$notify({
-					title: t('workspace', 'Error to rename space'),
-					text: t('workspace', 'Your Workspace name must not contain the following characters: [ ~ < > { } | ; . : , ! ? \' @ # $ + ( ) % \\\\ ^ = / & * ]'),
-					type: 'error',
-					duration: 6000,
-				})
-			}
 		},
 		// Sets a space's quota
 		setSpaceQuota(quota) {
@@ -260,9 +258,6 @@ export default {
 		},
 		toggleShowSelectUsersModal() {
 			this.showSelectUsersModal = !this.showSelectUsersModal
-		},
-		toggleShowDelWorkspaceModal() {
-			this.showDelWorkspaceModal = !this.showDelWorkspaceModal
 		},
 		updateColor(e) {
 			const spacename = this.$route.params.space
