+++ src/services/groupfoldersService.js
@@ -43,7 +43,6 @@ export function getAll() {
 			if (resp.data.ocs.meta.status === 'ok') {
 				return resp.data.ocs.data
 			}
-			throw new Error()
 		})
 		.catch(error => {
 			throw new BadGetError('Error to get all spaces', error.reason)
@@ -109,7 +108,7 @@ export function formatUsers(space) {
 /**
  * @param {string} spaceName it's the name of space to check
  * @param {object} vueInstance it's the instance of vue
- * @return {boolean}
+ * @return {Promise}
  * @throws {CheckGroupfolderNameExistError}
  */
 export async function checkGroupfolderNameExist(spaceName, vueInstance = undefined) {
@@ -204,7 +203,7 @@ export function addGroupToManageACLForGroupfolder(folderId, gid, vueInstance) {
 		})
 		.catch(error => {
 			if (typeof (vueInstance) !== 'undefined') {
-			showNotificationError('Error to add group as manager acl', 'Impossible to add the Space Manager group in Manage ACL groupfolder')
+				showNotificationError('Error to add group as manager acl', 'Impossible to add the Space Manager group in Manage ACL groupfolder')
 			}
 			console.error('Impossible to add the Space Manager group in Manage ACL groupfolder', error)
 			throw new AddGroupToManageACLForGroupfolderError('Error to add the Space Manager group in manage ACL groupfolder')
@@ -303,10 +302,9 @@ export function destroy(workspace) {
  *
  * @param {object} workspace it's the object relative to workspace
  * @param {string} newSpaceName it's the new name for the workspace
- * @param {object} vueInstance it's an instance from Vue
  * @return {Promise}
  */
-export function rename(workspace, newSpaceName, vueInstance = undefined) {
+export function rename(workspace, newSpaceName) {
 	// Response format to return
 	const respFormat = {
 		data: {},
@@ -314,6 +312,12 @@ export function rename(workspace, newSpaceName, vueInstance = undefined) {
 	respFormat.data.statuscode = 500
 	respFormat.data.message = 'Rename the space is impossible.'
 
+	if (!checkGroupfolderNameExist(workspace.name)) {
+		respFormat.data.statuscode = 409
+		respFormat.data.message = 'The space name already exist. We cannot rename with this name.'
+		console.error('The groupfolder name already exist. Please, choose another name to rename your space.')
+		return respFormat
+	}
 	newSpaceName = deleteBlankSpacename(newSpaceName)
 	// Update space side
 	const workspaceUpdated = axios.patch(generateUrl('/apps/workspace/api/space/rename'),
